Changes to accommodate v1.2
