+++ package.yaml
@@ -25,8 +25,7 @@ dependencies:
 - containers >= 0.5 && < 0.7
 - contravariant < 2
 - mtl >= 2.0.1.0 && < 3.0.0.0
-- stm >= 2 && < 3
-- polysemy >= 1.1.0.0
+- polysemy >= 1.2.0.0
 - polysemy-plugin >= 0.2
 - random >= 1.1 && <1.2
 - reflection >= 2.1.4 && < 3.0.0
@@ -75,7 +74,7 @@ tests:
     build-tools:
     - hspec-discover  >= 2.0
     dependencies:
-    - polysemy >= 1.1.0.0
+    - polysemy >= 1.2.0.0
     - polysemy-zoo
     - hspec >= 2.6.0 && < 3
